Fix: curPage has no default
All but one instance of `populateNiaiSection` call it with
nothing passed as the second argument, so a default should be set.

Fixed condition checking against "itemInfo"
which is not a possible value for the `on` field.

Fixed one instance where an empty string
was explicitly passed as the argument for `curPage`.